+++ player.cpp
@@ -69,7 +69,7 @@ namespace Tmpl8
 		if (fmod(invincibility, 0.5) < 0.25)
 		{
 			screen->CircleShadow({ pos.x - 5, pos.y + 5 }, 35, 0.7f);
-			screen->CircleFull(pos, 0, static_cast<int>(size.x / 2), 0x64b4ff);
+			screen->CircleFull(pos, 0, size.x / 2.0f, 0x64b4ff);
 			screen->CircleFull(pos, 20, 25, 0x5df5ff);
 		}
 	}
