+++ powerup.cpp
@@ -46,13 +46,13 @@ namespace Tmpl8
         if (active) {
             scale = 1.0f + whiteFlashTimer / 6.0f; // scale up when hit
             screen->BarShadow({ getTopLeft().x - 5, getTopLeft().y + 5 }, { getBottomRight().x - 5, getBottomRight().y + 5 }, 10.0f, 0.2f);
-        	const float barWidth = lifetime * 10;
+            const float barWidth = lifetime * 10;
             screen->BarShadow({ pos.x - barWidth / 2 * scale - 5, getBottomRight().y * scale + 15 }, { pos.x + barWidth / 2 - 5, getBottomRight().y + 25 }, 10.0f, 0.2f);
-        	screen->Bar(pos.x - barWidth / 2, getBottomRight().y + 10, pos.x + barWidth / 2, getBottomRight().y + 20, 0xffffff); // lifetime
-        	screen->Bar(pos - size.x / 2 * scale, pos + size.x / 2 * scale, 0xffb500); // body
+            screen->Bar(pos.x - barWidth / 2, getBottomRight().y + 10, pos.x + barWidth / 2, getBottomRight().y + 20, 0xffffff); // lifetime
+            screen->Bar(pos - size.x / 2 * scale, pos + size.x / 2 * scale, 0xffb500); // body
             screen->BoxThick(pos - size.x / 2 * scale, pos + size.x / 2 * scale, 5, 0xffd700); // outline
-            if (whiteFlashTimer > 0.0f) screen->Bar(pos - size.x / 2 * scale- 5, pos + size.x / 2 * scale + 5, 0xffffff, whiteFlashTimer); // white flash when hit
-            screen->PrintScaled("?", pos.x - 15, pos.y - 12, 5, 5, 0xffffff); // question mark
+            if (whiteFlashTimer > 0.0f) screen->Bar(pos - size.x / 2 * scale - 5, pos + size.x / 2 * scale + 5, 0xffffff, whiteFlashTimer); // white flash when hit
+            screen->PrintScaled("?", static_cast<int>(pos.x - 15), static_cast<int>(pos.y - 12), 5, 5, 0xffffff); // question mark
         }
 	}
 };
